Drop dependency on nix
Rewrite the vfio backend to rely only on libc, and swap the dependency
on nix (licensed as "MIT") by a dependency on libc (licensed as "MIT OR
Apache-2.0").

This gets us to a point where both our own code and all of our (non-dev)
dependencies are licensed as "MIT OR Apache-2.0".

Signed-off-by: Alberto Faria <afaria@redhat.com>